+++ duka-client/registration/inputHandlers/accountNumberInputHandler.js
@@ -45,10 +45,16 @@ module.exports = {
                     var registeredClient = registerClient(dcr_duka_client);
                     if(registeredClient) {
                         var messageToClient = getMessage('thanks_for_registering', {'$account_number': registeredClient.AccountNumber}, lang);
-                        project.sendMessage({
-                            content: messageToClient, 
-                            to_number: dcr_duka_client.phoneNumber
-                        });
+                        project.sendMulti({ messages: [
+                            {
+                                content: messageToClient, 
+                                to_number: dcr_duka_client.phoneNumber
+                            },
+                            {
+                                content: messageToClient, 
+                                to_number: contact.phone_number
+                            }
+                        ]});
                         console.log('duka client registered. Sent message: ' + messageToClient + ' to: ' + JSON.stringify(dcr_duka_client));
                         if(state.vars.dcr_credit > 0) {
                             global.sayText(getMessage('outstanding_balance', {'$balance': state.vars.dcr_credit}, lang));