+++ duka-client/registration/inputHandlers/confirmFirstSecondNameInputHandler.test.js
@@ -56,6 +56,7 @@ describe.each(['en-ke', 'sw'])('confirm first and second name using (%s)', (lang
         state.vars.duka_client_nid = '12345678';
         state.vars.duka_client_phone_number = '0788765453';
         state.vars.duka_client_invoice_id = 'axdf87';
+        contact.phone_number = '0788334455';
         registerClient.mockReturnValueOnce({AccountNumber: '12324565'});
         const rowMock = {save: jest.fn()};
         const tableMock = {createRow: jest.fn()};
@@ -63,15 +64,19 @@ describe.each(['en-ke', 'sw'])('confirm first and second name using (%s)', (lang
         jest.spyOn(project, 'getOrCreateDataTable').mockReturnValue(tableMock);
         confirmHandler(1);
         const message = {
-            'en-ke': {'content': 
-            'Thank you for registering with OAF. Your Account Number is 12324565. The Duka team will help you complete your loan.',
-            'to_number': '0788765453'},
-            'sw': {
-                'content': 
-            'Asante kwa kusajili na OAF. Nambari yako ya Akaunti ni 12324565. Timu ya Duka itakusaidia kumaliza mkopo wako',
-                'to_number': '0788765453'}
+            'en-ke': 'Thank you for registering with OAF. Your Account Number is 12324565. The Duka team will help you complete your loan.',
+            'sw': 'Asante kwa kusajili na OAF. Nambari yako ya Akaunti ni 12324565. Timu ya Duka itakusaidia kumaliza mkopo wako'
         };
-        expect(project.sendMessage).toHaveBeenCalledWith(message[lang]);
+        expect(project.scheduleMessage).toHaveBeenCalledWith({
+            content: message[lang], 
+            to_number: '0788334455',
+            start_time_offset: 0
+        });
+        expect(project.scheduleMessage).toHaveBeenCalledWith({
+            content: message[lang], 
+            to_number: '0788765453',
+            start_time_offset: 15
+        });
         expect(tableMock.createRow).toHaveBeenCalledWith({'vars': {'account_number': '12324565', 'invoice_id': 'axdf87', 'phone_number': '0788765453'}});
     });